replace shuttles
